'remove' need for block xor in the second pass of romix by doing it in place in chunkmix. idea from solar designer, but had to work out how to do it without requiring multiple copies of the mix function, and without slowing anything down noticeably

+++ scrypt-jane-speed.c
@@ -81,7 +81,7 @@ int main() {
 	topbit = 0;
 	for (i = cpuflags; i != 0; i >>= 1)
 		topbit++;
-	topbit = (1 << topbit);
+	topbit = ((size_t)1 << topbit);
 
 	while (1) {
 	#if defined(SCRYPT_CHOOSE_COMPILETIME)
